+++ src/core/public/chrome/nav_links/to_nav_link.ts
@@ -32,7 +32,7 @@ import {
   PublicAppInfo,
   AppNavLinkStatus,
   AppStatus,
-  WorkspaceAccessibility,
+  WorkspaceAvailability,
 } from '../../application';
 import { IBasePath } from '../../http';
 import { NavLinkWrapper } from './nav_link';
@@ -41,7 +41,7 @@ import { appendAppPath } from '../../application/utils';
 export function toNavLink(app: PublicAppInfo, basePath: IBasePath): NavLinkWrapper {
   const useAppStatus = app.navLinkStatus === AppNavLinkStatus.default;
   let relativeBaseUrl = basePath.prepend(app.appRoute!);
-  if (app.workspaceAccessibility === WorkspaceAccessibility.NO) {
+  if (app.workspaceAvailability === WorkspaceAvailability.outOfWorkspace) {
     relativeBaseUrl = basePath.prepend(app.appRoute!, { withoutClientBasePath: true });
   }
   const url = relativeToAbsolute(appendAppPath(relativeBaseUrl, app.defaultPath));