+++ src/plugins/workspace/public/plugin.ts
@@ -16,7 +16,7 @@ import {
   AppUpdater,
   AppStatus,
   PublicAppInfo,
-  WorkspaceAccessibility,
+  WorkspaceAvailability,
 } from '../../../core/public';
 import {
   WORKSPACE_FATAL_ERROR_APP_ID,
@@ -200,7 +200,7 @@ export class WorkspacePlugin implements Plugin<{}, {}, WorkspacePluginSetupDeps>
         const { renderCreatorApp } = await import('./application');
         return mountWorkspaceApp(params, renderCreatorApp);
       },
-      workspaceAccessibility: WorkspaceAccessibility.NO,
+      workspaceAvailability: WorkspaceAvailability.outOfWorkspace,
     });
 
     // update
@@ -246,7 +246,7 @@ export class WorkspacePlugin implements Plugin<{}, {}, WorkspacePluginSetupDeps>
         const { renderListApp } = await import('./application');
         return mountWorkspaceApp(params, renderListApp);
       },
-      workspaceAccessibility: WorkspaceAccessibility.NO,
+      workspaceAvailability: WorkspaceAvailability.outOfWorkspace,
     });
 
     /**